+++ NEWS.md
@@ -3,6 +3,7 @@
 * Restored `sig_ml` minimal value to the `estimate_parameters_gaussian()` function to prevent divide by zero errors in the `marginal_p_gaussian()` function. However, now the minimal value is instead set to the machine floating point minimum value. As a result, `biocorex()` may sometimes produce negative `tcs`. This is an indication that the data is non-gaussian and such results should not be trusted. This might occur, for example, when binary data is fit with a gaussian marginal description.
 * Minor code optimization to speed up calculating the discrete and bernoulli
 * Summary methods added
+* Experimental option logpx_method added. This allows use of an alternative method of calculating log_p_xi.
 
 # rcorex 0.2.2
 * Minor code optimization to speed up calculating the gaussian marginals